Relocate Debug namespace to reduce duplication (#31524)
